requirements.txt
autopep8==2.1.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
googlesearch-python==1.2.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
pycodestyle==2.11.1
python-docx==1.1.0
requests==2.31.0
soupsieve==2.5
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
